requirements.txt
crewai==0.11.2
python-dotenv>=1.0.0
flask>=3.0.0
requests>=2.31.0
